backend/requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg[binary]
google-adk
python-dotenv
python-docx
PyPDF2